[ENGA3-618]: Refactoring classes responsible for refunds
